requirements.txt
# Shared Python environment for the generative AI workshop.
#
# Use Python 3.11.
# This file is intended to be enough on its own:
#
#   pip install -r requirements.txt
#
# PyTorch is included explicitly so the VM template can be provisioned in one step.

accelerate>=1.7,<2
diffusers @ git+https://github.com/huggingface/diffusers.git@46bd0057308a6b3f270d235fef97d5d958433967
gradio>=5.33,<6
huggingface-hub>=0.32,<1
kokoro>=0.9,<1
matplotlib>=3.10,<4
numpy>=2.4,<3
ollama>=0.5,<0.6
openai-whisper>=20240930
Pillow>=11,<12
safetensors>=0.5,<0.6
scipy>=1.15,<2
sentencepiece>=0.2,<0.3
soundfile>=0.13,<0.14
torch>=2.10,<2.11
transformers>=4.52,<5
